Moving serialization out of Redis messenger into its own module
Modifying Messenger trait and Redis Messenger to operate on byte
slices rather than Solana/Geyser types.
