+++ messenger/src/lib.rs
@@ -7,17 +7,16 @@ pub trait Messenger {
     where
         Self: Sized;
 
-    fn send_account(&self, account: &ReplicaAccountInfo, slot: u64, is_startup: bool)
-        -> Result<()>;
-    fn send_slot_status(&self, slot: u64, parent: Option<u64>, status: SlotStatus) -> Result<()>;
-    fn send_transaction(
-        &mut self,
-        transaction_info: &ReplicaTransactionInfo,
-        slot: u64,
-    ) -> Result<()>;
-    fn send_block(&mut self, block_info: &ReplicaBlockInfo) -> Result<()>;
+    fn send_account(&mut self, bytes: &[u8]) -> Result<()>;
+    fn send_slot_status(&mut self, bytes: &[u8]) -> Result<()>;
+    fn send_transaction(&mut self, bytes: &[u8]) -> Result<()>;
+    fn send_block(&mut self, bytes: &[u8]) -> Result<()>;
+
     fn recv_account(&self) -> Result<()>;
     fn recv_slot_status(&self) -> Result<()>;
     fn recv_transaction(&self) -> Result<()>;
     fn recv_block(&self) -> Result<()>;
 }
+
+//fn send<T: MessengerKey>(&self, buffer: &[u8]) -> Result<()>;
+//fn recv<'a, T: MessengerKey>(&self, buffer: &'a [u8]) -> Result<&'a [u8]>;
